Deprecated obsolete functionality from the Scene
